Stats: use float64 for Histogram counts

+++ stats/buckets_test.go
@@ -195,7 +195,7 @@ func TestHistogram(t *testing.T) {
 			}
 			So(h.Size(), ShouldEqual, 1000)
 			So(h.Buckets().N, ShouldEqual, 10)
-			So(h.Counts(), ShouldResemble, []uint{
+			So(h.Counts(), ShouldResemble, []float64{
 				100, 100, 100, 100, 100, 100, 100, 100, 100, 100})
 			So(h.Count(5), ShouldEqual, 100)
 			So(h.Count(11), ShouldEqual, 0)
@@ -280,7 +280,7 @@ func TestHistogram(t *testing.T) {
 			So(h.Buckets().N, ShouldEqual, 9)
 			So(h.Buckets().Bounds, ShouldResemble, []float64{
 				-100.0, -10.0, -1.0, -0.1, -0.01, 0.01, 0.1, 1.0, 10.0, 100.0})
-			So(h.Counts(), ShouldResemble, []uint{
+			So(h.Counts(), ShouldResemble, []float64{
 				90, 9, 1, 0, 1, 0, 0, 9, 90})
 			So(testutil.RoundFixed(h.Mean(), 3), ShouldEqual, -0.5)       // actual: -0.5
 			So(testutil.Round(h.MAD(), 4), ShouldEqual, 50.0)             // actual: 50.0
